src/models/__init__.py
""" 
    import all models for sqlalchemy relational mapping
    We Don't use this file for importing our models except alembic

 """

from src.user.model import *
from src.auth.model import *
from src.cafe.model import *

__all__ = (
    "Base",

    "User", "Profile",
    "RevokedToken",

    "Cafe", "CafeComment",
)