requirements.txt
aioredis==2.0.1
alembic==1.17.2
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
asyncpg==0.31.0
billiard==4.2.4
celery==5.6.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.124.0
GeoAlchemy2==0.18.1
greenlet==3.3.0
h11==0.16.0
idna==3.11
kombu==5.6.2
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
passlib==1.7.4
prompt_toolkit==3.0.52
psycopg2==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
PyNaCl==1.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
redis==7.1.0
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
uvicorn==0.40.0
vine==5.1.0
wcwidth==0.2.14
